feat: update income transaction logic
