+++ document-models/real-world-assets/real-world-assets.json
@@ -13,7 +13,7 @@
             "changeLog": [],
             "state": {
                 "global": {
-                    "schema": "type RealWorldAssetsState {\n    accounts: [Account!]!\n    principalLenderAccountId: ID!\n    spvs: [Spv!]!\n    serviceProviderFeeTypes: [ServiceProviderFeeType!]!\n    fixedIncomeTypes: [FixedIncomeType!]!\n    portfolio: [Asset!]!\n    transactions: [GroupTransaction!]!\n}\n\nunion Asset = FixedIncome | Cash\n\ntype FixedIncome {\n    id: ID!\n type: AssetType!    fixedIncomeTypeId: ID!\n    name: String!\n    spvId: ID!\n    maturity: DateTime!\n    purchaseDate: DateTime!\n    notional: Float!\n    assetProceeds: Float!\n    purchaseProceeds: Float!\n    salesProceeds: Float!\n    purchasePrice: Float!\n    totalDiscount: Float!\n    realizedSurplus: Float!\n    ISIN: String\n    CUSIP: String\n    coupon: Float\n}\n\ntype Cash {\n    id: ID!\n  type: AssetType!  spvId: ID!\n    currency: String!\n    balance: Float!\n}\n\ntype FixedIncomeType {\n    id: ID!\n    name: String!\n}\n\ntype Spv {\n    id: ID!\n    name: String!\n}\n\nenum AssetType { Cash FixedIncome }\n\ntype BaseTransaction {\n    id: ID!\n assetType: AssetType!\n    assetId: ID!\n    amount: Float!\n    entryTime: DateTime!\n    tradeTime: DateTime\n    settlementTime: DateTime\n    txRef: String\n    accountId: ID\n    counterPartyAccountId: ID\n}\n\ntype Account {\n    id: ID!\n    reference: String!\n    label: String\n}\n\ntype TransactionFee {\n    id: ID!\n    serviceProviderFeeTypeId: ID!\n    amount: Float!\n}\n\ntype GroupTransaction {\n    id: ID!\n    type: GroupTransactionType!\n    entryTime: DateTime!\n    unitPrice: Float\n    cashBalanceChange: Float!\n    fees: [TransactionFee!]\n    cashTransaction: BaseTransaction!\n    fixedIncomeTransaction: BaseTransaction\n}\n\nenum GroupTransactionType {\n    PrincipalDraw\n    PrincipalReturn\n    AssetPurchase\n    AssetSale\n InterestIncome\n    InterestPayment\n  FeesIncome\n  FeesPayment\n}\n\ntype ServiceProviderFeeType {\n    id: ID!\n    name: String!\n    feeType: String!\n    accountId: ID!\n}\n\nscalar DateTime",
+                    "schema": "type RealWorldAssetsState {\n    accounts: [Account!]!\n    principalLenderAccountId: ID!\n    spvs: [Spv!]!\n    serviceProviderFeeTypes: [ServiceProviderFeeType!]!\n    fixedIncomeTypes: [FixedIncomeType!]!\n    portfolio: [Asset!]!\n    transactions: [GroupTransaction!]!\n}\n\nunion Asset = FixedIncome | Cash\n\ntype FixedIncome {\n    id: ID!\n type: AssetType!    fixedIncomeTypeId: ID!\n    name: String!\n    spvId: ID!\n    maturity: DateTime!\n    purchaseDate: DateTime!\n    notional: Float!\n    assetProceeds: Float!\n    purchaseProceeds: Float!\n    salesProceeds: Float!\n    purchasePrice: Float!\n    totalDiscount: Float!\n    realizedSurplus: Float!\n    ISIN: String\n    CUSIP: String\n    coupon: Float\n}\n\ntype Cash {\n    id: ID!\n  type: AssetType!  spvId: ID!\n    currency: String!\n    balance: Float!\n}\n\ntype FixedIncomeType {\n    id: ID!\n    name: String!\n}\n\ntype Spv {\n    id: ID!\n    name: String!\n}\n\nenum AssetType { Cash FixedIncome }\n\ntype BaseTransaction {\n    id: ID!\n assetType: AssetType!\n    assetId: ID!\n    amount: Float!\n    entryTime: DateTime!\n    tradeTime: DateTime\n    settlementTime: DateTime\n    txRef: String\n    accountId: ID\n    counterPartyAccountId: ID\n}\n\ntype Account {\n    id: ID!\n    reference: String!\n    label: String\n}\n\ntype TransactionFee {\n    id: ID!\n    serviceProviderFeeTypeId: ID!\n    amount: Float!\n}\n\ntype GroupTransaction {\n    id: ID!\n    type: GroupTransactionType!\n    entryTime: DateTime!\n    unitPrice: Float\n    cashBalanceChange: Float!\n    fees: [TransactionFee!]\n    cashTransaction: BaseTransaction!\n    fixedIncomeTransaction: BaseTransaction\n serviceProviderFeeTypeId: ID \n}\n\nenum GroupTransactionType {\n    PrincipalDraw\n    PrincipalReturn\n    AssetPurchase\n    AssetSale\n InterestIncome\n    InterestPayment\n  FeesIncome\n  FeesPayment\n}\n\ntype ServiceProviderFeeType {\n    id: ID!\n    name: String!\n    feeType: String!\n    accountId: ID!\n}\n\nscalar DateTime",
                     "initialValue": "{\"accounts\":[{\"id\":\"principal-lender-account-id\",\"reference\":\"0x\",\"label\":\"Principal Lender\"}],\"principalLenderAccountId\":\"principal-lender-account-id\",\"spvs\":[],\"serviceProviderFeeTypes\":[],\"fixedIncomeTypes\":[],\"portfolio\":[{\"id\":\"principal-asset-id\",\"type\":\"Cash\",\"spvId\":\"1\",\"currency\":\"USD\",\"balance\":0}],\"transactions\":[]}",
                     "examples": []
                 },
@@ -245,7 +245,7 @@
                             "id": "T3gWYNoqdT/tPrczgdpaCcQz5BI=",
                             "name": "CREATE_GROUP_TRANSACTION",
                             "description": "",
-                            "schema": "input CreateGroupTransactionInput {\n    id: ID!\n    type: GroupTransactionType!\n    entryTime: DateTime!\n fees: [TransactionFee!]\n cashBalanceChange: Float! unitPrice: Float   cashTransaction: BaseTransaction!\n    fixedIncomeTransaction: BaseTransaction\n}",
+                            "schema": "input CreateGroupTransactionInput {\n    id: ID!\n    type: GroupTransactionType!\n    entryTime: DateTime!\n fees: [TransactionFee!]\n unitPrice: Float   cashTransaction: BaseTransaction!\n    fixedIncomeTransaction: BaseTransaction\n serviceProviderFeeTypeId: ID \n}",
                             "template": "",
                             "reducer": "",
                             "errors": [],
@@ -256,7 +256,7 @@
                             "id": "Wg49korUNLaDgD8nhiOYd88S+6g=",
                             "name": "EDIT_GROUP_TRANSACTION",
                             "description": "",
-                            "schema": "input CreateGroupTransactionInput {\n    id: ID!\n    type: GroupTransactionType\n    entryTime: DateTime\n cashBalanceChange: Float! unitPrice: Float  cashTransaction: BaseTransaction!\n    fixedIncomeTransaction: BaseTransaction\n}",
+                            "schema": "input EditGroupTransactionInput {\n    id: ID!\n    type: GroupTransactionType\n    entryTime: DateTime\n unitPrice: Float  cashTransaction: BaseTransaction!\n    fixedIncomeTransaction: BaseTransaction\n serviceProviderFeeTypeId: ID \n}",
                             "template": "",
                             "reducer": "",
                             "errors": [],
