+++ document-models/real-world-assets/schema.graphql
@@ -87,6 +87,7 @@ type GroupTransaction {
     fees: [TransactionFee!]
     cashTransaction: BaseTransaction!
     fixedIncomeTransaction: BaseTransaction
+    serviceProviderFeeTypeId: ID
 }
 
 enum GroupTransactionType {
@@ -237,21 +238,21 @@ input CreateGroupTransactionInput {
     id: ID!
     type: GroupTransactionType!
     entryTime: DateTime!
-    cashBalanceChange: Float!
     cashTransaction: BaseTransaction!
     fees: [TransactionFee!]
     unitPrice: Float
     fixedIncomeTransaction: BaseTransaction
+    serviceProviderFeeTypeId: ID
 }
 
 input EditGroupTransactionInput {
     id: ID!
     type: GroupTransactionType
     entryTime: DateTime
-    cashBalanceChange: Float
     unitPrice: Float
     cashTransaction: BaseTransactionInput
     fixedIncomeTransaction: BaseTransactionInput
+    serviceProviderFeeTypeId: ID
 }
 
 input DeleteGroupTransactionInput {
