+++ document-models/real-world-assets/src/_utils/calculations.ts
@@ -1,8 +1,14 @@
 import {
     GroupTransaction,
     GroupTransactionType,
+    InputMaybe,
+    TransactionFee,
 } from 'document-models/real-world-assets';
-import { ASSET_PURCHASE, ASSET_SALE } from '../constants';
+import {
+    ASSET_PURCHASE,
+    ASSET_SALE,
+    cashTransactionSignByTransactionType,
+} from '../constants';
 
 /**
  * Compute derived fields for fixed income assets
@@ -250,4 +256,20 @@ export function roundToNearestDay(date: Date) {
     }
 
     return roundedDate;
+}
+
+export function calculateCashBalanceChange(
+    transactionType: InputMaybe<GroupTransactionType>,
+    cashAmount: InputMaybe<number>,
+    fees: InputMaybe<TransactionFee[]>,
+) {
+    if (!cashAmount || !transactionType) return 0;
+
+    const sign = cashTransactionSignByTransactionType[transactionType];
+
+    const feeAmounts = fees?.map(fee => fee.amount).filter(Boolean) ?? [];
+
+    const totalFees = feeAmounts.reduce((acc, fee) => acc + fee, 0);
+
+    return cashAmount * sign - totalFees;
 }
