+++ document-models/real-world-assets/src/_utils/validators/assets.ts
@@ -1,0 +1,88 @@
+import { InputMaybe } from 'document-model/document-model';
+import { Cash } from 'document-models/real-world-assets/gen';
+import { Asset, FixedIncome, RealWorldAssetsState } from '../../..';
+import { dateValidator, numberValidator } from './zod';
+
+export function isFixedIncomeAsset(
+    asset: Asset | undefined | null,
+): asset is FixedIncome {
+    if (!asset) return false;
+    if (asset.type === 'FixedIncome') return true;
+    if ('fixedIncomeId' in asset) return true;
+    return false;
+}
+
+export function isCashAsset(asset: Asset | undefined | null): asset is Cash {
+    if (!asset) return false;
+    if (asset.type === 'Cash') return true;
+    if ('currency' in asset) return true;
+    return false;
+}
+
+export function validateFixedIncomeAsset(
+    state: RealWorldAssetsState,
+    asset: InputMaybe<FixedIncome>,
+) {
+    if (!asset) return;
+
+    if (
+        asset.fixedIncomeTypeId &&
+        !state.fixedIncomeTypes.find(
+            fixedIncomeType => fixedIncomeType.id === asset.fixedIncomeTypeId,
+        )
+    ) {
+        throw new Error(
+            `Fixed income type with id ${asset.fixedIncomeTypeId} does not exist!`,
+        );
+    }
+    if (asset.spvId && !state.spvs.find(spv => spv.id === asset.spvId)) {
+        throw new Error(`SPV with id ${asset.spvId} does not exist!`);
+    }
+    if (asset.maturity && !dateValidator.safeParse(asset.maturity).success) {
+        throw new Error(`Maturity must be a valid date`);
+    }
+}
+
+export function validateFixedIncomeAssetDerivedFields(
+    asset: Partial<FixedIncome>,
+) {
+    if (
+        asset.purchaseDate &&
+        !dateValidator.safeParse(asset.purchaseDate).success
+    ) {
+        throw new Error(`Purchase date must be a valid date`);
+    }
+    if (asset.notional && !numberValidator.safeParse(asset.notional).success) {
+        throw new Error(`Notional must be a number`);
+    }
+    if (
+        asset.purchaseProceeds &&
+        !numberValidator.safeParse(asset.purchaseProceeds).success
+    ) {
+        throw new Error(`Purchase proceeds must be a number`);
+    }
+    if (
+        asset.purchasePrice &&
+        !numberValidator.safeParse(asset.purchasePrice).success
+    ) {
+        throw new Error(`Purchase price must be a number`);
+    }
+    if (
+        asset.totalDiscount &&
+        !numberValidator.safeParse(asset.totalDiscount).success
+    ) {
+        throw new Error(`Total discount must be a number`);
+    }
+    if (
+        asset.salesProceeds &&
+        !numberValidator.safeParse(asset.salesProceeds).success
+    ) {
+        throw new Error(`Annualized yield must be a number`);
+    }
+    if (
+        asset.realizedSurplus &&
+        !numberValidator.safeParse(asset.realizedSurplus).success
+    ) {
+        throw new Error(`Realized surplus must be a number`);
+    }
+}
