+++ document-models/real-world-assets/src/_utils/validators/zod.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod';
+
+export const dateValidator = z.coerce.date();
+
+export const numberValidator = z.number();
