+++ document-models/real-world-assets/src/reducers/transactions.ts
@@ -4,12 +4,12 @@
  * - delete the file and run the code generator again to have it reset
  */
 
+import { copy } from 'copy-anything';
 import {
     Cash,
+    calculateCashBalanceChange,
     isCashAsset,
     makeFixedIncomeAssetWithDerivedFields,
-    validateCashTransaction,
-    validateFixedIncomeTransaction,
     validateGroupTransaction,
     validateTransactionFee,
     validateTransactionFees,
@@ -25,12 +25,19 @@ export const reducer: RealWorldAssetsTransactionsOperations = {
 
         const type = action.input.type;
         const entryTime = action.input.entryTime;
-        const cashBalanceChange = action.input.cashBalanceChange;
         const unitPrice = action.input.unitPrice ?? null;
         const fees = action.input.fees ?? null;
+        const serviceProviderFeeTypeId =
+            action.input.serviceProviderFeeTypeId ?? null;
         let cashTransaction = action.input.cashTransaction;
         let fixedIncomeTransaction =
             action.input.fixedIncomeTransaction ?? null;
+        const cashBalanceChange = calculateCashBalanceChange(
+            type,
+            cashTransaction.amount,
+            fees,
+        );
+
         cashTransaction = {
             ...cashTransaction,
             entryTime,
@@ -52,9 +59,10 @@ export const reducer: RealWorldAssetsTransactionsOperations = {
             fees,
             cashTransaction,
             fixedIncomeTransaction,
+            serviceProviderFeeTypeId,
         };
 
-        validateGroupTransaction(newGroupTransaction, state);
+        validateGroupTransaction(state, newGroupTransaction);
 
         state.transactions.push(newGroupTransaction);
 
@@ -89,100 +97,118 @@ export const reducer: RealWorldAssetsTransactionsOperations = {
             throw new Error('Group transaction must have an id');
         }
 
-        const transaction = state.transactions.find(
+        const oldTransaction = state.transactions.find(
             transaction => transaction.id === action.input.id,
         );
 
-        if (!transaction) {
+        if (!oldTransaction) {
             throw new Error(
                 `Group transaction with id ${action.input.id} does not exist!`,
             );
         }
 
-        const oldCashBalanceChange = transaction.cashBalanceChange;
-        const newCashBalanceChange = action.input.cashBalanceChange;
+        // make a copy so that we can retain the old transaction values for updating related assets
+        const newTransaction = copy(oldTransaction);
 
-        const oldFixedIncomeAssetId =
-            transaction.fixedIncomeTransaction?.assetId;
-        const newFixedIncomeAssetId =
-            action.input.fixedIncomeTransaction?.assetId;
+        const {
+            type,
+            entryTime,
+            unitPrice,
+            serviceProviderFeeTypeId,
+            cashTransaction,
+            fixedIncomeTransaction,
+        } = action.input;
 
-        if (action.input.type) {
-            transaction.type = action.input.type;
+        if (type) {
+            newTransaction.type = type;
         }
 
-        if (action.input.entryTime) {
-            transaction.entryTime = action.input.entryTime;
-            transaction.cashTransaction.entryTime = action.input.entryTime;
-            if (transaction.fixedIncomeTransaction) {
-                transaction.fixedIncomeTransaction.entryTime =
-                    action.input.entryTime;
-            }
+        if (entryTime) {
+            newTransaction.entryTime = entryTime;
+            newTransaction.cashTransaction.entryTime = entryTime;
         }
 
-        if (
-            action.input.fixedIncomeTransaction?.amount &&
-            transaction.fixedIncomeTransaction
-        ) {
-            transaction.fixedIncomeTransaction.amount =
-                action.input.fixedIncomeTransaction.amount;
-            validateFixedIncomeTransaction(
-                state,
-                transaction.fixedIncomeTransaction,
+        if (unitPrice) {
+            newTransaction.unitPrice = unitPrice;
+        }
+
+        if (serviceProviderFeeTypeId) {
+            newTransaction.serviceProviderFeeTypeId = serviceProviderFeeTypeId;
+        }
+
+        if (cashTransaction?.amount) {
+            newTransaction.cashTransaction.amount = cashTransaction.amount;
+            newTransaction.cashBalanceChange = calculateCashBalanceChange(
+                newTransaction.type,
+                newTransaction.cashTransaction.amount,
+                newTransaction.fees,
             );
         }
 
-        if (action.input.cashTransaction?.amount) {
-            transaction.cashTransaction.amount =
-                action.input.cashTransaction.amount;
-            validateCashTransaction(state, transaction.cashTransaction);
+        if (newTransaction.fixedIncomeTransaction) {
+            if (fixedIncomeTransaction?.amount) {
+                newTransaction.fixedIncomeTransaction.amount =
+                    fixedIncomeTransaction.amount;
+            }
+
+            if (fixedIncomeTransaction?.assetId) {
+                newTransaction.fixedIncomeTransaction.assetId =
+                    fixedIncomeTransaction.assetId;
+            }
+
+            newTransaction.fixedIncomeTransaction.entryTime =
+                newTransaction.entryTime;
         }
 
-        if (action.input.unitPrice) {
-            transaction.unitPrice = action.input.unitPrice;
-        }
+        // first validate and update the transaction
+        validateGroupTransaction(state, newTransaction);
 
-        if (newFixedIncomeAssetId && transaction.fixedIncomeTransaction) {
-            transaction.fixedIncomeTransaction.assetId = newFixedIncomeAssetId;
-        }
+        state.transactions = state.transactions.map(t =>
+            t.id === newTransaction.id ? newTransaction : t,
+        );
 
-        if (newCashBalanceChange) {
-            transaction.cashBalanceChange = newCashBalanceChange;
+        // if successfully updated transaction, also update related assets
 
+        // if cash amount has changed, update the cash asset in state
+        if (cashTransaction?.amount) {
             const cashAsset = state.portfolio.find(a => isCashAsset(a)) as Cash;
 
-            cashAsset.balance += newCashBalanceChange - oldCashBalanceChange;
+            cashAsset.balance +=
+                newTransaction.cashBalanceChange -
+                oldTransaction.cashBalanceChange;
 
             state.portfolio = state.portfolio.map(a =>
                 a.id === cashAsset.id ? cashAsset : a,
             );
         }
 
-        state.transactions = state.transactions.map(t =>
-            t.id === transaction.id ? transaction : t,
-        );
-
-        if (oldFixedIncomeAssetId) {
+        // if the existing transaction had a fixed income asset, update that asset to reflect the changes
+        if (oldTransaction.fixedIncomeTransaction?.assetId) {
             const updatedOldFixedIncomeAsset =
                 makeFixedIncomeAssetWithDerivedFields(
                     state,
-                    oldFixedIncomeAssetId,
+                    oldTransaction.fixedIncomeTransaction.assetId,
                 );
 
             state.portfolio = state.portfolio.map(a =>
-                a.id === oldFixedIncomeAssetId ? updatedOldFixedIncomeAsset : a,
+                a.id === updatedOldFixedIncomeAsset.id
+                    ? updatedOldFixedIncomeAsset
+                    : a,
             );
         }
 
-        if (newFixedIncomeAssetId) {
+        // if the new transaction has a fixed income asset, update that asset to have the effects of the transaction
+        if (fixedIncomeTransaction?.assetId) {
             const updatedNewFixedIncomeAsset =
                 makeFixedIncomeAssetWithDerivedFields(
                     state,
-                    newFixedIncomeAssetId,
+                    fixedIncomeTransaction.assetId,
                 );
 
             state.portfolio = state.portfolio.map(a =>
-                a.id === newFixedIncomeAssetId ? updatedNewFixedIncomeAsset : a,
+                a.id === updatedNewFixedIncomeAsset.id
+                    ? updatedNewFixedIncomeAsset
+                    : a,
             );
         }
     },
@@ -219,13 +245,7 @@ export const reducer: RealWorldAssetsTransactionsOperations = {
             a.id === fixedIncomeAssetId ? updatedFixedIncomeAsset : a,
         );
 
-        const cashAssetId = transactionToRemove.cashTransaction.assetId;
-
-        if (!cashAssetId) return;
-
-        const cashAsset = state.portfolio.find(
-            a => a.id === cashAssetId,
-        ) as Cash;
+        const cashAsset = state.portfolio.find(a => isCashAsset(a)) as Cash;
 
         const updatedCashAsset = {
             ...cashAsset,
@@ -233,7 +253,7 @@ export const reducer: RealWorldAssetsTransactionsOperations = {
         };
 
         state.portfolio = state.portfolio.map(a =>
-            a.id === cashAssetId ? updatedCashAsset : a,
+            a.id === cashAsset.id ? updatedCashAsset : a,
         );
     },
     addFeesToGroupTransactionOperation(state, action, dispatch) {
@@ -255,8 +275,25 @@ export const reducer: RealWorldAssetsTransactionsOperations = {
 
         transaction.fees.push(...action.input.fees);
 
+        transaction.cashBalanceChange = calculateCashBalanceChange(
+            transaction.type,
+            transaction.cashTransaction.amount,
+            transaction.fees,
+        );
+
         state.transactions = state.transactions.map(t =>
             t.id === action.input.id ? transaction : t,
+        );
+
+        const cashAsset = state.portfolio.find(a => isCashAsset(a)) as Cash;
+
+        const updatedCashAsset = {
+            ...cashAsset,
+            balance: cashAsset.balance + transaction.cashBalanceChange,
+        };
+
+        state.portfolio = state.portfolio.map(a =>
+            a.id === cashAsset.id ? updatedCashAsset : a,
         );
     },
     removeFeesFromGroupTransactionOperation(state, action, dispatch) {
@@ -279,8 +316,25 @@ export const reducer: RealWorldAssetsTransactionsOperations = {
             fee => !feeIdsToRemove?.includes(fee.id),
         );
 
+        transaction.cashBalanceChange = calculateCashBalanceChange(
+            transaction.type,
+            transaction.cashTransaction.amount,
+            transaction.fees,
+        );
+
         state.transactions = state.transactions.map(t =>
             t.id === id ? transaction : t,
+        );
+
+        const cashAsset = state.portfolio.find(a => isCashAsset(a)) as Cash;
+
+        const updatedCashAsset = {
+            ...cashAsset,
+            balance: cashAsset.balance + transaction.cashBalanceChange,
+        };
+
+        state.portfolio = state.portfolio.map(a =>
+            a.id === cashAsset.id ? updatedCashAsset : a,
         );
     },
     editGroupTransactionFeesOperation(state, action, dispatch) {
@@ -311,8 +365,25 @@ export const reducer: RealWorldAssetsTransactionsOperations = {
             })
             .filter(Boolean);
 
+        transaction.cashBalanceChange = calculateCashBalanceChange(
+            transaction.type,
+            transaction.cashTransaction.amount,
+            transaction.fees,
+        );
+
         state.transactions = state.transactions.map(t =>
             t.id === id ? transaction : t,
+        );
+
+        const cashAsset = state.portfolio.find(a => isCashAsset(a)) as Cash;
+
+        const updatedCashAsset = {
+            ...cashAsset,
+            balance: cashAsset.balance + transaction.cashBalanceChange,
+        };
+
+        state.portfolio = state.portfolio.map(a =>
+            a.id === cashAsset.id ? updatedCashAsset : a,
         );
     },
 };
